added deduplication via relations
